Struct xshell::Cmd

pub struct Cmd { /* fields omitted */ }

A command.

Implementations

impl Cmd

pub fn new(program: impl AsRef<Path>) -> Cmd

Creates a new Cmd that executes the given program.

pub fn arg(self, arg: impl AsRef<OsStr>) -> Cmd

Pushes an argument onto this Cmd.

pub fn args<I>(self, args: I) -> Cmd where
I: IntoIterator,
I::Item: AsRef<OsStr>,

Pushes the arguments onto this Cmd.

pub fn env<K, V>(self, key: K, val: V) -> Cmd where
K: AsRef<OsStr>,
V: AsRef<OsStr>,

Equivalent to std::process::Command::env.

pub fn envs<I, K, V>(self, vars: I) -> Cmd where
I: IntoIterator<Item = (K, V)>,
K: AsRef<OsStr>,
V: AsRef<OsStr>,

Equivalent to std::process::Command::envs.

Note: This does not replace the child process’s environment, unless you call Cmd::env_clear first.

pub fn env_remove<K>(self, key: K) -> Cmd where
K: AsRef<OsStr>,

Equivalent to std::process::Command::env_remove.

pub fn env_clear(self) -> Cmd

Equivalent to std::process::Command::env_clear.

Note that on Windows some environmental variables are required for process spawning. See https://github.com/rust-lang/rust/issues/31259.

pub fn ignore_stdout(self) -> Cmd

Returns a Cmd that will ignore the stdout stream. This is equivalent of attaching stdout to /dev/null.

pub fn ignore_stderr(self) -> Cmd

Returns a Cmd that will ignore the stderr stream. This is equivalent of attaching stderr to /dev/null.

pub fn ignore_status(self) -> Cmd

Returns a Cmd that ignores its exit status.

pub fn stdin(self, stdin: impl AsRef<[u8]>) -> Cmd

Returns a Cmd with the given stdin.

pub fn echo_cmd(self, echo: bool) -> Cmd

Returns a Cmd that echoes itself (or not) as specified.

pub fn secret(self, secret: bool) -> Cmd

Returns a Cmd that is secret (or not) as specified.

If a command is secret, it echoes <secret> instead of the program and its arguments.

pub fn read(self) -> Result<String>

Returns the stdout from running the command.

pub fn read_stderr(self) -> Result<String>

Returns the stderr from running the command.

pub fn output(self) -> Result<Output>

Returns a std::process::Output from running the command.

pub fn run(self) -> Result<()>

Runs the command.